fix bug button 'voir plus' in reservation view
